requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
pydantic
langchain-classic
faiss-cpu
langchain-text-splitters
pypdf



